Extracted the k8s runtime implementation into a crate
