+++ lib/k8s-runtime/src/state/evmap.rs
@@ -1,8 +1,8 @@
-//! A state implementation backed by [`evmap10`].
+//! A state implementation backed by [`evmap`].
 
-use crate::kubernetes::{debounce::Debounce, hash_value::HashValue};
+use crate::{debounce::Debounce, hash_value::HashValue};
 use async_trait::async_trait;
-use evmap10::WriteHandle;
+use evmap::WriteHandle;
 use futures::future::BoxFuture;
 use k8s_openapi::{apimachinery::pkg::apis::meta::v1::ObjectMeta, Metadata};
 use std::time::Duration;
@@ -10,6 +10,7 @@ use std::time::Duration;
 /// A [`WriteHandle`] wrapper that implements [`super::Write`].
 /// For use as a state writer implementation for
 /// [`crate::kubernetes::Reflector`].
+#[derive(Debug)]
 pub struct Writer<T>
 where
     T: Metadata<Ty = ObjectMeta> + Send,
@@ -128,7 +129,7 @@ fn kv<T: Metadata<Ty = ObjectMeta>>(object: T) -> Option<(String, Value<T>)> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::kubernetes::state::{MaintainedWrite, Write};
+    use crate::state::{MaintainedWrite, Write};
     use k8s_openapi::api::core::v1::Pod;
 
     fn make_pod(uid: &str) -> Pod {
@@ -151,7 +152,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_without_debounce() {
-        let (state_reader, state_writer) = evmap10::new();
+        let (state_reader, state_writer) = evmap::new();
         let mut state_writer = Writer::new(state_writer, None);
 
         assert_eq!(state_reader.is_empty(), true);
@@ -170,7 +171,7 @@ mod tests {
         // Due to https://github.com/tokio-rs/tokio/issues/2090 we're not
         // pausing the time.
 
-        let (state_reader, state_writer) = evmap10::new();
+        let (state_reader, state_writer) = evmap::new();
         let flush_debounce_timeout = Duration::from_millis(100);
         let mut state_writer = Writer::new(state_writer, Some(flush_debounce_timeout));
