+++ src/internal_events/kubernetes/mod.rs
@@ -2,8 +2,5 @@
 
 use super::InternalEvent;
 
-pub mod api_watcher;
 pub mod instrumenting_state;
 pub mod instrumenting_watcher;
-pub mod reflector;
-pub mod stream;
