+++ src/sources/kubernetes_logs/mod.rs
@@ -167,16 +167,16 @@ impl Source {
         let field_selector = format!("spec.nodeName={}", self_node_name);
         let label_selector = "vector.dev/exclude!=true".to_owned();
 
-        let watcher = k8s::api_watcher::ApiWatcher::new(client, Pod::watch_pod_for_all_namespaces);
+        let watcher = k8s_runtime::client::Watcher::new(client, Pod::watch_pod_for_all_namespaces);
         let watcher = k8s::instrumenting_watcher::InstrumentingWatcher::new(watcher);
         let (state_reader, state_writer) = evmap::new();
         let state_writer =
-            k8s::state::evmap::Writer::new(state_writer, Some(Duration::from_millis(10)));
-        let state_writer = k8s::state::instrumenting::Writer::new(state_writer);
+            k8s_runtime::state::evmap::Writer::new(state_writer, Some(Duration::from_millis(10)));
+        let state_writer = k8s::instrumenting_state::Writer::new(state_writer);
         let state_writer =
-            k8s::state::delayed_delete::Writer::new(state_writer, Duration::from_secs(60));
+            k8s_runtime::state::delayed_delete::Writer::new(state_writer, Duration::from_secs(60));
 
-        let mut reflector = k8s::reflector::Reflector::new(
+        let mut reflector = k8s_runtime::reflector::Reflector::new(
             watcher,
             state_writer,
             Some(field_selector),
